[feat] BindingFragment 생성
