+++ app/src/main/java/com/sopt/now/fragment/HomeFragment.kt
@@ -1,10 +1,7 @@
 package com.sopt.now.fragment
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.GridLayoutManager
 import com.sopt.now.databinding.FragmentHomeBinding
@@ -12,22 +9,10 @@ import com.sopt.now.friend.FriendAdapter
 import com.sopt.now.viewmodel.HomeViewModel
 import com.sopt.now.viewmodel.MainViewModel.Companion.LOGIN_INFO
 
-class HomeFragment() : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding: FragmentHomeBinding
-        get() = requireNotNull(_binding)
+class HomeFragment() : BindingFragment<FragmentHomeBinding>(FragmentHomeBinding::inflate) {
 
     private lateinit var friendAdapter: FriendAdapter
     private lateinit var viewModel: HomeViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -49,12 +34,6 @@ class HomeFragment() : Fragment() {
         viewModel.userInfo.observe(viewLifecycleOwner) {
             friendAdapter.setUser(it.data)
         }
-    }
-
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
 }
