+++ app/src/main/java/com/sopt/now/fragment/MyPageFragment.kt
@@ -2,11 +2,8 @@ package com.sopt.now.fragment
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.appcompat.app.AlertDialog
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.sopt.now.R
 import com.sopt.now.activity.ChangePwdActivity
@@ -16,22 +13,10 @@ import com.sopt.now.viewmodel.MainViewModel.Companion.LOGIN_INFO
 import com.sopt.now.viewmodel.MainViewModel.Companion.USER_INFO
 import com.sopt.now.viewmodel.MyPageViewModel
 
-class MyPageFragment() : Fragment() {
-    private var _binding: FragmentMyPageBinding? = null
-    private val binding: FragmentMyPageBinding
-        get() = requireNotNull(_binding)
+class MyPageFragment() : BindingFragment<FragmentMyPageBinding>(FragmentMyPageBinding::inflate) {
 
     private var userId: String? = null
     private lateinit var viewModel: MyPageViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentMyPageBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -84,11 +69,6 @@ class MyPageFragment() : Fragment() {
         val intent = Intent(requireContext(), LoginActivity::class.java)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
         activity?.startActivity(intent)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
 }
