+++ app/src/main/res/layout/activity_change_pwd.xml
@@ -45,7 +45,7 @@
         app:layout_constraintTop_toBottomOf="@id/tv_new_pwd" />
 
     <TextView
-        android:id="@+id/tv_check_pwd"
+        android:id="@+id/tv_new_pwd_verification"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="30dp"
@@ -54,13 +54,13 @@
         app:layout_constraintTop_toBottomOf="@id/et_new_pwd" />
 
     <EditText
-        android:id="@+id/et_check_pwd"
+        android:id="@+id/et_new_pwd_verification"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:hint="@string/previous_new_hint"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_check_pwd" />
+        app:layout_constraintTop_toBottomOf="@id/tv_new_pwd_verification" />
 
     <Button
         android:id="@+id/btn_change_pwd"
